updated scripts
made a check for filesystem to override any errors

+++ bashcom_lib/text.sh
@@ -13,49 +13,53 @@
 clear
 echo -e "=== Text Message ===\n"
 
-# Ask if we're searching through address book or not
-echo -ne "Text someone from address book? (y/n)\n: "
-read answer
-
-if [[ ("$answer" = "y") || ("$answer" = "Y") ]]
+# Check to see if address book exists
+if [ -f $BOOK ];
 then
-	# Ask for the name or number of person to text
-	echo -n "Enter name or number of person you want to text: "
-    read find
+    # Ask if we're searching through address book or not
+    echo -ne "Text someone from address book? (y/n)\n: "
+    read answer
 
-    # Get the number of results with that string in the file
-    results=$(cat $BOOK | grep -ci "$find")
-
-    # Check if it's greater than 1 or not
-    if [ "$results" -gt 1 ]
+    if [[ ("$answer" = "y") || ("$answer" = "Y") ]]
     then
-    	echo -e "\n$results result(s) found in address book:\n"
-    	cat $BOOK | grep -i $find > ./bashcom_lib/temp.txt
-    	(awk -F ":" '{printf "Entry %d:\n%s\n%s\n==========================\n\n", NR, $1, $2}' ./bashcom_lib/temp.txt)
+    	# Ask for the name or number of person to text
+    	echo -n "Enter name or number of person you want to text: "
+        read find
 
-    	# Ask user to select the user they want to text
-    	echo -ne "Enter entry number of the contact you want to text: "
-    	read linenum
+        # Get the number of results with that string in the file
+        results=$(cat $BOOK | grep -ci "$find")
 
-    	# Save the number to the phone variable
-     	PHONE=$(cat ./bashcom_lib/temp.txt | awk "NR==$linenum" | awk -F':' '{print $2}')
-     	rm ./BashText/temp.txt
-    else
-    	echo -e "\n$results result found in address book:"
-    	# Display the entry selected
-		cat $BOOK | grep -i $find
+        # Check if it's greater than 1 or not
+        if [ "$results" -gt 1 ]
+        then
+        	echo -e "\n$results result(s) found in address book:\n"
+        	cat $BOOK | grep -i $find > ./bashcom_lib/temp.txt
+        	(awk -F ":" '{printf "Entry %d:\n%s\n%s\n==========================\n\n", NR, $1, $2}' ./bashcom_lib/temp.txt)
 
-		echo -ne "\nText this person? (y/n): "
-		read isPerson
+        	# Ask user to select the user they want to text
+        	echo -ne "Enter entry number of the contact you want to text: "
+        	read linenum
 
-		if [[ ("$isPerson" = "y") || ("$isPerson" = "Y") ]]
-		then
-			# Save the number to the phone variable
-			PHONE=$(cat $BOOK | grep -i "$find" | awk -F':' '{print $2}')
-		else
-			echo -n "Enter phone number: "
-			read PHONE
-		fi
+        	# Save the number to the phone variable
+         	PHONE=$(cat ./bashcom_lib/temp.txt | awk "NR==$linenum" | awk -F':' '{print $2}')
+         	rm ./BashText/temp.txt
+        else
+        	echo -e "\n$results result found in address book:"
+        	# Display the entry selected
+    		cat $BOOK | grep -i $find
+
+    		echo -ne "\nText this person? (y/n): "
+    		read isPerson
+
+    		if [[ ("$isPerson" = "y") || ("$isPerson" = "Y") ]]
+    		then
+    			# Save the number to the phone variable
+    			PHONE=$(cat $BOOK | grep -i "$find" | awk -F':' '{print $2}')
+    		else
+    			echo -n "Enter phone number: "
+    			read PHONE
+    		fi
+        fi
     fi
 else
 	echo -n "Enter phone number: "
